refactor(projects): split the command module by subgroup
`src/commands/projects.ts` was a single 890-line file, and the work
queued behind this commit adds three more subgroups to it (status
updates, relations, status administration). Growing one file to ~2000
lines would make it the largest command module in the repo by a wide
margin and would bury the domain registration under the CRUD.

Adopt the layout `src/commands/initiatives/` already uses: `index.ts`
owns `PROJECTS_META`, the domain command, and the `usage` subcommand;
`entity.ts` owns everything registered on it. Each new subgroup lands as
its own sibling file wired from `index.ts`, so a reader looking for
"what does `projects` expose" reads one short file instead of scanning
for `.command(` calls.

Discussions stay in `entity.ts` rather than moving to a file of their
own, matching `initiatives/entity.ts` — they are registered flat on the
domain, not as a subgroup, and splitting them would diverge from the
module this layout is copied from.

Pure move plus import-path adjustment; no command, flag, or output
changed.
